+++ src/main/kotlin/cloud/glitchdev/rfu/feature/fishing/ReindrakeAlert.kt
@@ -1,7 +1,7 @@
 package cloud.glitchdev.rfu.feature.fishing
 
 import cloud.glitchdev.rfu.config.categories.JerryFishing
-import cloud.glitchdev.rfu.events.managers.ChatEvents.registerChatEvent
+import cloud.glitchdev.rfu.events.managers.ChatEvents.registerGameEvent
 import cloud.glitchdev.rfu.feature.Feature
 import cloud.glitchdev.rfu.feature.RFUFeature
 import cloud.glitchdev.rfu.utils.Title
@@ -11,8 +11,8 @@ object ReindrakeAlert : Feature {
     private val REINDRAKE_REGEX = Regex("^WOAH! (?:\\[.*?\\] )?(\\w+) summoned (a|TWO) Reindrakes? from the depths!$")
 
     override fun onInitialize() {
-        registerChatEvent(REINDRAKE_REGEX) { _, match ->
-            if (match == null || !JerryFishing.reindrakeAlert) return@registerChatEvent
+        registerGameEvent(REINDRAKE_REGEX) { _, _, match ->
+            if (match == null || !JerryFishing.reindrakeAlert) return@registerGameEvent
 
             val count = match.groupValues[2]